pyproject.toml
[project]
name = "album-genre-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.9",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "spotipy>=2.26.0",
    "tqdm>=4.67.3",
]
